Rename priority enum files
More generic, so we can add further enums here.
