+++ test-db-manager/src/types/index.ts
@@ -1,3 +1,3 @@
 export * from './common';
 export * from './inputs';
-export * from './priority';
+export * from './enums';
