+++ ui/src/components/forms/common/PriorityForm.tsx
@@ -2,7 +2,7 @@ import { useFormContext } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 import { z } from 'zod';
 import { Column, Row, Visible } from '../../../layoutComponents';
-import { Priority } from '../../../types/Priority';
+import { Priority } from '../../../types/enums';
 import { SimpleButton } from '../../../uiComponents';
 
 export const priorityFormSchema = z.object({
