+++ ui/src/components/forms/route/TemplateRouteSelector.tsx
@@ -2,7 +2,7 @@ import { DateTime } from 'luxon';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Column, Row } from '../../../layoutComponents';
-import { Priority } from '../../../types/Priority';
+import { Priority } from '../../../types/enums';
 import { SimpleButton, ValueFn } from '../../../uiComponents';
 import { ChooseRouteDropdown } from './ChooseRouteDropdown';
