+++ ui/src/components/routes-and-lines/line-details/LineValidityPeriod.tsx
@@ -3,7 +3,7 @@ import { RouteLine } from '../../../generated/graphql';
 import { mapPriorityToUiName } from '../../../i18n/uiNameMappings';
 import { Row } from '../../../layoutComponents';
 import { DateLike, mapToShortDate } from '../../../time';
-import { Priority } from '../../../types/Priority';
+import { Priority } from '../../../types/enums';
 
 const testIds = {
   validityPeriod: 'LineValidityPeriod::validityPeriod',
