+++ ui/src/types/enums.ts
